organized my code
