+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/OpModes/Autonomous/BluePaths/RightBluePark.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode.OpModes.Autonomous.BluePaths;
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
